backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
pandas
numpy
joblib
psycopg2-binary
gunicorn
python-dotenv